--- a/api/image.go
+++ b/api/image.go
@@ -8,7 +8,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -177,7 +176,16 @@ func BuildDockerImage(ctx context.Context, cli MeliAPiClient, dc *DockerContaine
 			originalErr: err,
 			newErr:      errors.New(" :unable to build docker image")}
 	}
-	io.Copy(dc.LogMedium, imageBuildResponse.Body)
+
+	var imgProg ImageProgress
+	scanner := bufio.NewScanner(imageBuildResponse.Body)
+	for scanner.Scan() {
+		_ = json.Unmarshal(scanner.Bytes(), &imgProg)
+		fmt.Fprintln(dc.LogMedium, dc.ServiceName, "::", imgProg.Status, imgProg.Progress)
+	}
+	if err := scanner.Err(); err != nil {
+		fmt.Println(" :unable to log output for image", imageName, err)
+	}
 
 	imageBuildResponse.Body.Close()
 	return imageName, nil