//new benchmarks:
```bash
BenchmarkGetAuth-32                 	  100000	     15688 ns/op
BenchmarkCreateContainer-32         	  300000	      5346 ns/op
BenchmarkContainerStart-32          	10000000	       197 ns/op
BenchmarkContainerLogs-32           	  100000	     14867 ns/op
BenchmarkFormatLabels-32            	 1000000	      1391 ns/op
BenchmarkFormatPorts-32             	 2000000	       712 ns/op
BenchmarkFormatServiceVolumes-32    	 1000000	      1206 ns/op
BenchmarkFormatImageName-32         	 1000000	      1924 ns/op
BenchmarkPullDockerImage-32         	  100000	     17024 ns/op
BenchmarkBuildDockerImage-32        	  200000	     11427 ns/op
BenchmarkGetNetwork-32              	 5000000	       280 ns/op
BenchmarkConnectNetwork-32          	 3000000	       444 ns/op
BenchmarkCreateDockerVolume-32      	  500000	      3265 ns/op
```
// old benchmarks(ie master branch)
```bash
BenchmarkGetAuth-32                 	  100000	     13401 ns/op
BenchmarkCreateContainer-32         	  200000	      6000 ns/op
BenchmarkContainerStart-32          	10000000	       198 ns/op
BenchmarkContainerLogs-32           	  200000	      6539 ns/op
BenchmarkFormatLabels-32            	 1000000	      1312 ns/op
BenchmarkFormatPorts-32             	 2000000	       798 ns/op
BenchmarkFormatServiceVolumes-32    	 1000000	      1218 ns/op
BenchmarkFormatImageName-32         	 1000000	      1977 ns/op
BenchmarkPullDockerImage-32         	  200000	      7174 ns/op
BenchmarkBuildDockerImage-32        	  200000	     12336 ns/op
BenchmarkGetNetwork-32              	 5000000	       306 ns/op
BenchmarkConnectNetwork-32          	 3000000	       471 ns/op
BenchmarkCreateDockerVolume-32      	  500000	      3487 ns/op
```