@komuW 2017-11-10
// new time to `meli -up`
`real	0m5.374s`

// old time to `meli -up`
`real	0m5.308s`